requirements.txt
selenium
webdriver_manager
yt_dlp
openai-whisper
transformers
openai
requests==2.31.0
google-cloud-storage
pandas==2.2.1
plotly==5.19.0
streamlit==1.32.0
python-dateutil==2.8.2
google-api-python-client
